Implementacion del resto de validaciones de back end

+++ src/controllers/userController.js
@@ -3,28 +3,41 @@ const fs = require('fs');
 const path = require('path');
 const { validationResult } = require('express-validator');
 const db = require('../database/models');
+const Op = db.Sequelize.Op;
 const bcrypt = require('bcryptjs');
 
 const userController = {
     register: (req, res) => res.render('register'),
     store: (req,res) => {
         let errors = validationResult(req);
-
+        
         if (errors.isEmpty()) {
-            db.User.create({
-                email: req.body.email,
-                username: req.body.username, 
-                admin: req.body.category,
-                password: bcrypt.hashSync(req.body.password, 10),
-                image: req.file ? req.file.filename : 'default-image.png'
+            db.User.findOne ({
+                where: {
+                    username: req.body.username
+                }
             }).then( user => {
-                req.session.user = { 
-                    id: user.id,
-                    username: user.username,
-                    admin: user.admin,
-                    image: user.image
-                };
-                res.redirect('/');
+                if (user) {
+                    let errors = {
+                        username: { msg: 'El nombre de usuario ya se encuentra en uso' }
+                    }
+                    return res.render('register', { errors: errors, old: req.body });
+                }
+                db.User.create({
+                    email: req.body.email,
+                    username: req.body.username, 
+                    admin: req.body.category,
+                    password: bcrypt.hashSync(req.body.password, 10),
+                    image: req.file ? req.file.filename : 'default-image.png'
+                }).then( user => {
+                    req.session.user = { 
+                        id: user.id,
+                        username: user.username,
+                        admin: user.admin,
+                        image: user.image
+                    };
+                    res.redirect('/');
+                });
             });
         }
         else {
@@ -58,7 +71,10 @@ const userController = {
                     res.redirect('/');
                 }
                 else {
-                    res.render('login');
+                    let errors = {
+                        credentials: { msg: 'Usuario o contraseña incorrectos' }
+                    }
+                    res.render('login', { errors: errors, old: req.body });
                 }
             });
         }
@@ -77,27 +93,40 @@ const userController = {
         let errors = validationResult(req);
 
         if (errors.isEmpty()) {
-            db.User.update(
-                {
-                    email: req.body.email,
-                    username: req.body.username, 
-                    admin: req.body.category,
-                    image: req.file ? req.file.filename : req.session.user.image 
-                },
-                {
-                    where: { id: req.session.user.id }
-                })
-                    .then( () => {
-                        db.User.findByPk (req.session.user.id)
-                            .then( user => {
-                                if (req.file && req.session.user.image != 'default-image.png') {
-                                    fs.unlinkSync(path.resolve(__dirname, `../../public/img/users/${req.session.user.image}`));
-                                };
-                                req.session.user.username = user.username;
-                                req.session.user.admin = user.admin;
-                                req.session.user.image = user.image;
-                                res.redirect('/account');
-                            });
+            db.User.findOne ({
+                where: {
+                    username: req.body.username,
+                    id: { [Op.not]: req.session.user.id }
+                }
+            }).then( user => {
+                if (user) {
+                    let errors = {
+                        username: { msg: 'El nombre de usuario ya se encuentra en uso' }
+                    }
+                    return res.render('userEdit', { errors: errors, old: req.body });
+                }
+                db.User.update(
+                    {
+                        email: req.body.email,
+                        username: req.body.username, 
+                        admin: req.body.category,
+                        image: req.file ? req.file.filename : req.session.user.image 
+                    },
+                    {
+                        where: { id: req.session.user.id }
+                    })
+                        .then( () => {
+                            db.User.findByPk (req.session.user.id)
+                                .then( user => {
+                                    if (req.file && req.session.user.image != 'default-image.png') {
+                                        fs.unlinkSync(path.resolve(__dirname, `../../public/img/users/${req.session.user.image}`));
+                                    };
+                                    req.session.user.username = user.username;
+                                    req.session.user.admin = user.admin;
+                                    req.session.user.image = user.image;
+                                    res.redirect('/account');
+                                });
+                        });
                     });
         }
         else {
